+++ lib/data/product_list.dart
@@ -1,12 +1,10 @@
 import 'package:hw4/models/product.dart';
 
-List<Product> productList = [
-  Product("Apple", 3800, " ", "iphone 14"),
-  product1,
-];
+List<Product> productList = [];
 
 Product product1 = Product("Apple", 3800, "lib/image/img_1.png", "iphone 14");
 Product product2 = Product("Apple", 2000, "lib/image/img_2.png", "iphone xr");
 Product product3 = Product("Apple", 1800, "lib/image/img_3.png", "iphone x");
 Product product4 = Product("Samsung", 3600, "lib/image/img_4.png", "ultra 23");
-Product product5 = Product("huawei", 5000, "lib/image/img_5.png", "matebook d14");
+Product product5 = Product("huawei", 1000, "lib/image/img_6.png", "vivo y56");
+Product product6 = Product("huawei", 1200, "lib/image/img_5.png", "oppo A16k");
