+++ lib/main.dart
@@ -10,6 +10,8 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(home: HomeScreen());
+    return const MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: HomeScreen());
   }
 }
