+++ lib/screens/home_screen.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/material.dart';
 
 import '../data/product_list.dart';
-import '../models/product.dart';
+import '../models/item_caed.dart';
 import 'list_cart.dart';
-import 'product_detail_screen.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key});
@@ -73,6 +72,9 @@ class _HomeScreenState extends State<HomeScreen> {
                                 ItemCard(
                                   product: product5,
                                 ),
+                                ItemCard(
+                                  product: product6,
+                                ),
                               ],
                             ),
                           ),
@@ -89,47 +91,3 @@ class _HomeScreenState extends State<HomeScreen> {
     );
   }
 }
-
-class ItemCard extends StatelessWidget {
-  final Product product;
-
-  const ItemCard({
-    Key? key,
-    required this.product,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        //push the product detail screen
-        Navigator.push(
-            context,
-            MaterialPageRoute(
-                builder: (context) => ProductDetailScreen(product: product)));
-      },
-      child: Card(
-        //set the color of the card
-        color: Color.fromARGB(255, 200, 195, 195),
-        child: Column(
-          children: [
-            Image.asset(product.image),
-            Text(
-              product.price.toString(),
-              style: const TextStyle(fontSize: 20),
-            ),
-            Text(
-              product.name,
-              style: const TextStyle(fontSize: 15),
-            ),
-            Text(
-              product.brand,
-              style: const TextStyle(
-                  fontSize: 10, color: Color.fromRGBO(41, 41, 43, 1)),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
