+++ lib/screens/list_cart.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../models/item_row.dart';
 import '../models/product.dart';
 
 class ListCart extends StatefulWidget {
@@ -41,7 +42,7 @@ class _ListCartState extends State<ListCart> {
             Padding(
               padding: const EdgeInsets.all(16.0),
               child: Text(
-                'Total Price: \$${totalPrice.toStringAsFixed(2)}',
+                'Total Price: ${totalPrice.toStringAsFixed(2)}',
                 style: const TextStyle(
                     color: Colors.white,
                     fontSize: 30,
@@ -74,34 +75,6 @@ class _ListCartState extends State<ListCart> {
           ],
         ),
       ),
-    );
-  }
-}
-
-class ItemRow extends StatelessWidget {
-  final Product product;
-
-  const ItemRow({
-    Key? key,
-    required this.product,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        SizedBox(height: 50, width: 50, child: Image.asset(product.image)),
-        const SizedBox(width: 35),
-        Text(
-          product.name,
-          style: const TextStyle(fontSize: 20, color: Colors.white),
-        ),
-        const SizedBox(width: 35),
-        Text(
-          product.price.toString(),
-          style: const TextStyle(fontSize: 15, color: Colors.white),
-        ),
-      ],
     );
   }
 }
